[Networking] Add the NWFramer implementation. (#7338)
Uncommented the sources and update some mistakes after following the
sample provided by Apple. Initially tests were going to be added but
they resulted to be to flacky and would make the CI red too often to be
adding value.

Porting the sample will ensure that it works are the bindings are not
broken.