[InputMethodKit] Add ignored framework.
The InputMethodKit framework was ignored because we did not have a
sample, but that is not longer the case. An apple sample can be found
here: https://developer.apple.com/library/archive/samplecode/NumberInput_IMKit_Sample/Introduction/Intro.html

Once the bindings are landed we can work in the sample.

The framework was never added therefore we are not changing any API from
the old code since it was never released.

Fixes: https://github.com/xamarin/xamarin-macios/issues/7487