[eventkitui] Remove caching in `EKUIBundle.UIBundle`
The existing property implementation use `=` (not `=>`) which means that
the call is executed only once (in the `.cctor`).

```csharp
public static NSBundle UIBundle { get; } = Runtime.GetNSObject<NSBundle> (EventKitUIBundle ());
```

While a shorter syntax (good) it becomes easy to miss the difference
while reviewing pull requests (bad) or later when reading the code.

Also the documentation on `EventKitUIBundle` [1] is unclear. It mention
it's based on _user preferences_ which can be interpreted as something
that change at runtime. This would mean the returned pointer can be
different between calls and that we can't cache it from the static
constructor (since we would only return the original one).

So the correct (safe) code would be:

```csharp
public static NSBundle UIBundle { get; } => Runtime.GetNSObject<NSBundle> (EventKitUIBundle ());
```

but that's still hard to review so the older, self explanatory, syntax is
used.

[1] https://developer.apple.com/documentation/eventkitui/2866511-eventkituibundle?language=objc