Use a socket over a stream.
This is an attempt to fix https://github.com/xamarin/maccore/issues/827
by moving away from the stream and too a socket. If we got a 0, the user
disconnected and we do not write anything to the log. In theory this
should stop writing as soon as the client is out and fix the issue.