[Harness] Use a tcp tunnel to communicate when running iOS tests.
The tcp connection is very fragile and has been giving a lot of problems
in the labs. With the landing of
https://github.com/xamarin/maccore/pull/2158 and
https://github.com/spouliot/Touch.Unit/pull/56 we can use a tcp tunnel
that will communicate with the device via the usb cable. This means that
we will not longer have communication issues with iOS devices in the
labs.

The tcp tunnel is only used on iOS devices, not simulator, not tvOS or
watchOS.