[src] Stop using assembly identity while computing stret-ness.
Using assembly identity to verify that a type is the expected one becomes
complicated with .NET 5, because types move around a lot between assemblies.
There are also type forwarders all over the place which doesn't make things
better. So instead use (only) string comparisons on type names to determine
which types we're dealing with.

The only time this change can become a problem is if someone decides to
implement their own version of the built-in System types, but that would be a
very bad idea for a number of other reasons.