[Harness] When on device, check if we got a tcp exception on a crash.
Some crashes are reported as tcp connection issues because they happen
before the app haad the chance to write anything in the log. In that
case, we checked for the presence of the file, and if not present we
decided it was a tcp issue when it is not the case.

In this commit, we check for the tcp erorr message in the main_log so
that we are certain that the issue was with the connection and not
anyother.

Testingis simple, ran tests without the phone being part of the same
network and test with a branch that has a crash. For example the one in
https://github.com/xamarin/xamarin-macios/pull/8009 hash 83240612e8c0b2f7ed20547b92dd42a112b24d7c